split store to multiple files
